[refactor] move package Location DispatcherServletTest
