+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -26,8 +26,8 @@ public class AnnotationHandlerMappingTest {
         MockHttpServletRequest request = new MockHttpServletRequest("GET", "/users");
         request.setParameter("userId", user.getUserId());
         MockHttpServletResponse response = new MockHttpServletResponse();
-        HandlerExecution execution = handlerMapping.getHandler(request);
-        execution.handle(request, response);
+        Object execution = handlerMapping.getHandler(request);
+        ((HandlerExecution) execution).handle(request, response);
 
         assertThat(request.getAttribute("user")).isEqualTo(user);
     }
@@ -39,7 +39,7 @@ public class AnnotationHandlerMappingTest {
         request.setParameter("name", user.getName());
         request.setParameter("email", user.getEmail());
         MockHttpServletResponse response = new MockHttpServletResponse();
-        HandlerExecution execution = handlerMapping.getHandler(request);
-        execution.handle(request, response);
+        Object execution = handlerMapping.getHandler(request);
+        ((HandlerExecution) execution).handle(request, response);
     }
 }
